circuit-breaker/requirements.txt
pybreaker==0.6.0
paho-mqtt==1.5.1